new version of small_memory_tree

+++ durak_computer_controlled_opponent/database.cxx
@@ -76,18 +76,19 @@ gameStateAsString (std::tuple<std::vector<uint8_t>, std::vector<uint8_t> > const
   return vectorToString (std::get<0> (cards)) + ";" + vectorToString (std::get<1> (cards)) + ";" + std::to_string (magic_enum::enum_integer (trump));
 }
 
-small_memory_tree::SmallMemoryTree<std::tuple<Action, Result>, uint16_t>
+small_memory_tree::SmallMemoryTree<std::tuple<Action, Result>, uint8_t>
 binaryToSmallMemoryTree (std::string movesAndResultAsBinary)
 {
   auto ss = std::stringstream{ movesAndResultAsBinary };
   auto archive = cereal::BinaryInputArchiveWithContainingVectorSize{ ss };
-  auto smallMemoryTree = small_memory_tree::SmallMemoryTree<std::tuple<Action, Result>, uint16_t>{};
+  auto smallMemoryTree = small_memory_tree::SmallMemoryTree<std::tuple<Action, Result>, uint8_t>{};
   archive (smallMemoryTree);
   return smallMemoryTree;
 }
 std::string
-smallMemoryTreeToBinary (small_memory_tree::SmallMemoryTree<std::tuple<Action, Result>, uint16_t> const &smallMemoryTree)
+smallMemoryTreeToBinary (small_memory_tree::SmallMemoryTree<std::tuple<Action, Result>, uint8_t> const &smallMemoryTree)
 {
+  // TODO calculate children count per node and save this so we do not need to use 2 bytes and can use 1 byte instead
   auto ss = std::stringstream{};
   auto archive = cereal::BinaryOutputArchiveWithContainingVectorSize{ ss };
   archive (smallMemoryTree);
@@ -95,7 +96,7 @@ smallMemoryTreeToBinary (small_memory_tree::SmallMemoryTree<std::tuple<Action, R
 }
 
 void
-insertGameLookUp (std::filesystem::path const &databasePath, std::map<std::tuple<uint8_t, uint8_t>, std::array<std::map<std::tuple<std::vector<uint8_t>, std::vector<uint8_t> >, small_memory_tree::SmallMemoryTree<std::tuple<Action, Result>, uint16_t> >, 4> > const &gameLookup)
+insertGameLookUp (std::filesystem::path const &databasePath, std::map<std::tuple<uint8_t, uint8_t>, std::array<std::map<std::tuple<std::vector<uint8_t>, std::vector<uint8_t> >, small_memory_tree::SmallMemoryTree<std::tuple<Action, Result>, uint8_t> >, 4> > const &gameLookup)
 {
   soci::session sql (soci::sqlite3, databasePath.c_str ());
   soci::transaction tr (sql);
